requirements.txt
Pillow>=9.0.0
openai>=1.0.0
replicate>=0.11.0
python-dotenv>=0.19.0